integrator/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
httpx>=0.25.0
python-dotenv>=1.0.0
torch>=2.0.0
numpy>=1.24.0
